+++ pmd-jsp/src/main/java/net/sourceforge/pmd/lang/jsp/cpd/JSPTokenizer.java
@@ -9,14 +9,14 @@ import net.sourceforge.pmd.lang.TokenManager;
 import net.sourceforge.pmd.lang.ast.impl.javacc.CharStream;
 import net.sourceforge.pmd.lang.ast.impl.javacc.JavaccToken;
 import net.sourceforge.pmd.lang.document.TextDocument;
-import net.sourceforge.pmd.lang.jsp.ast.JspParser;
+import net.sourceforge.pmd.lang.jsp.ast.InternalApiBridge;
 import net.sourceforge.pmd.lang.jsp.ast.JspTokenKinds;
 
 public class JSPTokenizer extends JavaCCTokenizer {
 
     @Override
     protected TokenManager<JavaccToken> makeLexerImpl(TextDocument doc) {
-        return JspTokenKinds.newTokenManager(CharStream.create(doc, JspParser.getTokenBehavior()));
+        return JspTokenKinds.newTokenManager(CharStream.create(doc, InternalApiBridge.getJspTokenBehavior()));
     }
 
 }